x86: remove static declaration from multiboot_header definition
Newer versions of gcc fail to compile OpenBIOS giving a
"error: ‘multiboot_header’ defined but not used" message when the default
build with -Werror enabled is used.

Remove the static declaration to enable compilation to succeed, which
incidentally matches the other copies of multiboot.c in the OpenBIOS
source tree.

Signed-off-by: Mark Cave-Ayland <mark.cave-ayland@ilande.co.uk>