SPARC32: implement dvma_map_in() for DMA IOVA to phys translation
Rather than have dvma_alloc() store the result of the physical translation
in the specified location, use the fact that we now have contiguous physical
DMA memory to implement dvma_map_in() to perform the translation separately.

This allows us to remove the pointer to the physical address parameter to
dvma_alloc() and instead perform the translation in esp.c by calling
dvma_map_in() separately.

Signed-off-by: Mark Cave-Ayland <mark.cave-ayland@ilande.co.uk>