SPARC64: implement dma-* words
Implement the dma-* words in the root node using Forth, except for dma-alloc
and dma-sync which require C to use the in-built aligned memory allocator and
MMU flushing.

Note that as OpenBIOS doesn't enable the IOMMU for SPARC64 it is possible to
use the aligned memory allocator directly without any extra support code.

Signed-off-by: Mark Cave-Ayland <mark.cave-ayland@ilande.co.uk>