pci: add dma-* call-parent chain to all PCI devices and bridges
This enables disk drivers to call the dma-* words as required by the
specification, and in particular fixes the failing dma-alloc warnings
emitted by NetBSD PPC on boot.

Note that we also add these methods to any partition/file packages via a
new is-call-parent helper word to ensure that the dma-* words are
available at all levels of interposition.

Signed-off-by: Mark Cave-Ayland <mark.cave-ayland@ilande.co.uk>