.github/workflows: add main.yml for building OpenBIOS upon push
This is a GitHub push action that builds OpenBIOS for all of the currently
supported architectures (amd64, sparc32, sparc64, ppc and x86) and generates
an output zip file containing debug and release binaries.

The output zip file is stored both as a build artifact (which has a maximum
lifetime of 90 days) and for upstream OpenBIOS builds a "latest" release is
added to the repository.

Signed-off-by: Mark Cave-Ayland <mark.cave-ayland@ilande.co.uk>